ai-legislative-analyzer/backend/requirements.txt
fastapi
uvicorn
deep-translator>=1.11.4
langdetect>=1.0.9
sqlalchemy
psycopg2-binary
python-multipart
pdfplumber
sentence-transformers
chromadb
rank-bm25
openai
python-dotenv
pydantic
pydantic-settings
google-genai
pytesseract
opencv-python
feedparser